mapping; jwt; behavior; att entidades

+++ src/Core/Agropet.Application/UseCases/Lote/Commands/AtualizarLoteCommand.cs
@@ -7,7 +7,7 @@ public sealed record AtualizarLoteCommand : IRequest<Resposta>
 {
     public int Id { get; set; }
     public string? Numero { get; set; }
-    public double Quantidade { get; set; }
+    public int Quantidade { get; set; }
     public string? UnidadeComercial { get; set; }
     public decimal PrecoUnitarioCompra { get; set; }
     public DateTime? DataFabricacao { get; set; }
